+++ src/components/about-us/section-02/departments-name-list.js
@@ -1,16 +1,19 @@
-import colors from '../../../constants/colors'
-import { font } from '../constants/styles'
-import { gray } from './utils'
 import PropTypes from 'prop-types'
 import React, { PureComponent } from 'react'
 import styled from 'styled-components'
+// utils
+import { gray } from './utils'
+// constants
+import { font } from '../constants/styles'
+// @twreporter
+import { colorGrayscale } from '@twreporter/core/lib/constants/color'
 
 const DepartmentName = styled.div`
   p {
     font-size: 14px;
     font-weight: ${font.weight.bold};
     word-spacing: -1px;
-    color: ${props => (props.selected ? colors.black : gray.lightgray)};
+    color: ${props => (props.selected ? colorGrayscale.black : gray.lightgray)};
     width: max-content;
     margin-right: 20px;
   }